# public/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "GTM Skills API",
    "version": "1.0.0",
    "description": "The definitive API for AI-powered Go-To-Market prompts. Access 2,500+ battle-tested prompts for B2B sales and marketing.",
    "contact": {
      "name": "GTM Skills",
      "url": "https://gtm-skills.com",
      "email": "hello@gtm-skills.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "https://gtm-skills.com/api/v1",
      "description": "Production"
    }
  ],
  "paths": {
    "/prompts": {
      "get": {
        "summary": "List prompts",
        "description": "Retrieve a list of prompts with optional filtering",
        "operationId": "listPrompts",
        "tags": ["Prompts"],
        "parameters": [
          {
            "name": "category",
            "in": "query",
            "description": "Filter by category",
            "schema": {
              "type": "string",
              "example": "saas"
            }
          },
          {
            "name": "subcategory",
            "in": "query",
            "description": "Filter by subcategory",
            "schema": {
              "type": "string",
              "example": "outreach"
            }
          },
          {
            "name": "industry",
            "in": "query",
            "description": "Filter by industry",
            "schema": {
              "type": "string",
              "enum": ["saas", "financial-services", "healthcare", "manufacturing", "professional-services", "ecommerce", "real-estate", "education", "logistics", "energy", "government", "telecom", "media", "hospitality", "construction", "insurance"]
            }
          },
          {
            "name": "tags",
            "in": "query",
            "description": "Comma-separated list of tags to filter by",
            "schema": {
              "type": "string",
              "example": "cold-email,outreach"
            }
          },
          {
            "name": "difficulty",
            "in": "query",
            "description": "Filter by difficulty level",
            "schema": {
              "type": "string",
              "enum": ["beginner", "intermediate", "advanced"]
            }
          },
          {
            "name": "search",
            "in": "query",
            "description": "Search in title, description, and content",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Number of results (max 100)",
            "schema": {
              "type": "integer",
              "default": 20,
              "maximum": 100
            }
          },
          {
            "name": "offset",
            "in": "query",
            "description": "Pagination offset",
            "schema": {
              "type": "integer",
              "default": 0
            }
          },
          {
            "name": "sort",
            "in": "query",
            "description": "Sort order",
            "schema": {
              "type": "string",
              "enum": ["popular", "recent", "alphabetical"]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PromptsListResponse"
                }
              }
            }
          }
        }
      }
    },
    "/prompts/{id}": {
      "get": {
        "summary": "Get prompt by ID",
        "description": "Retrieve a single prompt by its ID",
        "operationId": "getPrompt",
        "tags": ["Prompts"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "The prompt ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PromptResponse"
                }
              }
            }
          },
          "404": {
            "description": "Prompt not found"
          }
        }
      }
    },
    "/prompts/recommend": {
      "get": {
        "summary": "Get prompt recommendations",
        "description": "Get contextual prompt recommendations based on deal stage, persona, industry, and context",
        "operationId": "recommendPrompts",
        "tags": ["Prompts"],
        "parameters": [
          {
            "name": "deal_stage",
            "in": "query",
            "description": "Current deal stage",
            "schema": {
              "type": "string",
              "enum": ["prospecting", "discovery", "demo", "proposal", "negotiation", "closing"]
            }
          },
          {
            "name": "persona",
            "in": "query",
            "description": "Buyer persona (e.g., 'VP Sales', 'CTO')",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "industry",
            "in": "query",
            "description": "Target industry",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "company_size",
            "in": "query",
            "description": "Target company size",
            "schema": {
              "type": "string",
              "enum": ["startup", "smb", "mid-market", "enterprise"]
            }
          },
          {
            "name": "context",
            "in": "query",
            "description": "Free-form context about the situation",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Number of recommendations (max 20)",
            "schema": {
              "type": "integer",
              "default": 5,
              "maximum": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecommendationsResponse"
                }
              }
            }
          },
          "400": {
            "description": "At least one context parameter is required"
          }
        }
      }
    },
    "/categories": {
      "get": {
        "summary": "List categories",
        "description": "List all available categories (industries, roles, workflows, methodologies)",
        "operationId": "listCategories",
        "tags": ["Categories"],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoriesResponse"
                }
              }
            }
          }
        }
      }
    },
    "/contributors": {
      "get": {
        "summary": "List contributors",
        "description": "Retrieve a list of approved contributors with their stats",
        "operationId": "listContributors",
        "tags": ["Contributors"],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "description": "Number of results (max 50)",
            "schema": {
              "type": "integer",
              "default": 20,
              "maximum": 50
            }
          },
          {
            "name": "offset",
            "in": "query",
            "description": "Pagination offset",
            "schema": {
              "type": "integer",
              "default": 0
            }
          },
          {
            "name": "featured",
            "in": "query",
            "description": "Filter to featured contributors only",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "description": "Sort order",
            "schema": {
              "type": "string",
              "enum": ["earnings", "prompts", "revenue", "recent"]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContributorsListResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Register as contributor",
        "description": "Submit an application to become a contributor",
        "operationId": "registerContributor",
        "tags": ["Contributors"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ContributorRegistration"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Registration submitted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContributorRegistrationResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request"
          }
        }
      }
    },
    "/contributors/{slug}": {
      "get": {
        "summary": "Get contributor by slug",
        "description": "Retrieve a single contributor's public profile",
        "operationId": "getContributor",
        "tags": ["Contributors"],
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "description": "The contributor's URL slug",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "stats",
            "in": "query",
            "description": "Include detailed statistics",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContributorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Contributor not found"
          }
        }
      }
    },
    "/contributors/leaderboard": {
      "get": {
        "summary": "Get contributor leaderboard",
        "description": "Get top contributors by earnings, revenue influenced, and prompts",
        "operationId": "getContributorLeaderboard",
        "tags": ["Contributors"],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContributorLeaderboardResponse"
                }
              }
            }
          }
        }
      }
    },
    "/attribution": {
      "post": {
        "summary": "Track attribution event",
        "description": "Track a contributor attribution event (visit, prompt view, copy, outcome)",
        "operationId": "trackAttribution",
        "tags": ["Attribution"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AttributionEvent"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Attribution tracked successfully"
          },
          "400": {
            "description": "Invalid request"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Prompt": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique prompt identifier"
          },
          "title": {
            "type": "string",
            "description": "Prompt title"
          },
          "description": {
            "type": "string",
            "description": "Brief description of the prompt"
          },
          "prompt": {
            "type": "string",
            "description": "The full prompt content"
          },
          "category": {
            "type": "string",
            "description": "Primary category"
          },
          "subcategory": {
            "type": "string",
            "description": "Subcategory"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Tags for filtering"
          },
          "difficulty": {
            "type": "string",
            "enum": ["beginner", "intermediate", "advanced"],
            "description": "Difficulty level"
          }
        }
      },
      "RecommendedPrompt": {
        "allOf": [
          {
            "$ref": "#/components/schemas/Prompt"
          },
          {
            "type": "object",
            "properties": {
              "relevance_score": {
                "type": "number",
                "description": "Relevance score (0-100)"
              },
              "relevance_reason": {
                "type": "string",
                "description": "Explanation of why this prompt is relevant"
              }
            }
          }
        ]
      },
      "PromptsListResponse": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Prompt"
            }
          },
          "pagination": {
            "type": "object",
            "properties": {
              "total": {
                "type": "integer"
              },
              "limit": {
                "type": "integer"
              },
              "offset": {
                "type": "integer"
              },
              "hasMore": {
                "type": "boolean"
              }
            }
          },
          "meta": {
            "type": "object",
            "properties": {
              "api_version": {
                "type": "string"
              },
              "documentation": {
                "type": "string"
              }
            }
          }
        }
      },
      "PromptResponse": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/Prompt"
          },
          "meta": {
            "type": "object"
          }
        }
      },
      "RecommendationsResponse": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RecommendedPrompt"
            }
          },
          "context": {
            "type": "object"
          },
          "meta": {
            "type": "object"
          }
        }
      },
      "CategoriesResponse": {
        "type": "object",
        "properties": {
          "data": {
            "type": "object",
            "properties": {
              "industries": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "slug": {
                      "type": "string"
                    },
                    "name": {
                      "type": "string"
                    },
                    "description": {
                      "type": "string"
                    },
                    "count": {
                      "type": "integer"
                    }
                  }
                }
              },
              "roles": {
                "type": "array",
                "items": {
                  "type": "object"
                }
              },
              "workflows": {
                "type": "array",
                "items": {
                  "type": "object"
                }
              },
              "methodologies": {
                "type": "array",
                "items": {
                  "type": "object"
                }
              }
            }
          },
          "meta": {
            "type": "object"
          }
        }
      },
      "Contributor": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique contributor identifier"
          },
          "name": {
            "type": "string",
            "description": "Contributor's display name"
          },
          "slug": {
            "type": "string",
            "description": "URL-friendly identifier"
          },
          "avatar_url": {
            "type": "string",
            "description": "Profile image URL"
          },
          "bio": {
            "type": "string",
            "description": "Contributor bio"
          },
          "twitter_handle": {
            "type": "string"
          },
          "linkedin_url": {
            "type": "string"
          },
          "total_prompts": {
            "type": "integer",
            "description": "Number of approved prompts"
          },
          "total_copies": {
            "type": "integer",
            "description": "Total prompt copies"
          },
          "total_outcomes": {
            "type": "integer",
            "description": "Total outcomes reported"
          },
          "total_revenue_influenced": {
            "type": "number",
            "description": "Total revenue influenced in USD"
          },
          "verified": {
            "type": "boolean",
            "description": "Whether contributor is verified"
          },
          "featured": {
            "type": "boolean",
            "description": "Whether contributor is featured"
          }
        }
      },
      "ContributorsListResponse": {
        "type": "object",
        "properties": {
          "contributors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Contributor"
            }
          },
          "total": {
            "type": "integer"
          },
          "limit": {
            "type": "integer"
          },
          "offset": {
            "type": "integer"
          }
        }
      },
      "ContributorResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/Contributor"
          },
          {
            "type": "object",
            "properties": {
              "github_handle": {
                "type": "string"
              },
              "website_url": {
                "type": "string"
              },
              "created_at": {
                "type": "string",
                "format": "date-time"
              },
              "stats": {
                "type": "object",
                "properties": {
                  "total_earnings": {
                    "type": "number"
                  },
                  "pending_payout": {
                    "type": "number"
                  },
                  "this_month_earnings": {
                    "type": "number"
                  },
                  "top_prompt": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string"
                      },
                      "title": {
                        "type": "string"
                      },
                      "copies": {
                        "type": "integer"
                      },
                      "revenue": {
                        "type": "number"
                      }
                    }
                  }
                }
              }
            }
          }
        ]
      },
      "ContributorRegistration": {
        "type": "object",
        "required": ["email", "name"],
        "properties": {
          "email": {
            "type": "string",
            "format": "email"
          },
          "name": {
            "type": "string"
          },
          "bio": {
            "type": "string"
          },
          "twitter_handle": {
            "type": "string"
          },
          "linkedin_url": {
            "type": "string"
          }
        }
      },
      "ContributorRegistrationResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "contributor": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string"
              },
              "name": {
                "type": "string"
              },
              "slug": {
                "type": "string"
              },
              "status": {
                "type": "string"
              }
            }
          },
          "message": {
            "type": "string"
          }
        }
      },
      "ContributorLeaderboardResponse": {
        "type": "object",
        "properties": {
          "top_by_earnings": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Contributor"
            }
          },
          "top_by_revenue": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Contributor"
            }
          },
          "top_by_prompts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Contributor"
            }
          }
        }
      },
      "AttributionEvent": {
        "type": "object",
        "required": ["contributor_id", "visitor_fingerprint", "event_type"],
        "properties": {
          "contributor_id": {
            "type": "string",
            "description": "ID of the contributor to credit"
          },
          "prompt_id": {
            "type": "string",
            "description": "ID of the prompt if applicable"
          },
          "visitor_fingerprint": {
            "type": "string",
            "description": "Unique visitor identifier"
          },
          "event_type": {
            "type": "string",
            "enum": ["visit", "prompt_view", "prompt_copy", "outcome_logged", "signup", "upgrade"],
            "description": "Type of attribution event"
          },
          "event_value": {
            "type": "number",
            "description": "Dollar value if applicable"
          },
          "referrer_url": {
            "type": "string"
          },
          "landing_page": {
            "type": "string"
          },
          "utm_source": {
            "type": "string"
          },
          "utm_medium": {
            "type": "string"
          },
          "utm_campaign": {
            "type": "string"
          }
        }
      }
    }
  }
}